requirements.txt
kafka-python==2.0.2
pandas==2.2.3
clickhouse-connect==0.8.12
python-dateutil==2.9.0.post0
